Сделал страницу 403

+++ articles/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
-
+from django.contrib.auth.decorators import login_required
 from articles import forms
 from articles.models import Post
 from django.http import HttpResponse, JsonResponse
@@ -27,18 +27,25 @@ def post(request, pk):
                    'prev_post': prev_post})
 
 
+@login_required(login_url='/users/sign_in')
 def post_create(request):
     form = forms.PostForm(request.POST or None, request.FILES)
     if request.method == 'POST' and form.is_valid():
-        form.save()
+        instance = form.save(commit=False)
+        instance.author = request.user
+        instance.save()
         return redirect('articles:home')
     return render(request, 'post_create.html', {'form': form})
 
 
+@login_required(login_url='/users/sign_in')
 def post_edit(request, pk):
-    article = Post.objects.get( pk=pk)
+    article = Post.objects.get(pk=pk)
     form = forms.PostForm(request.POST or None, request.FILES or None,
                           instance=article)
+    if article.author != request.user:
+        return redirect('users:forbidden')
+
     if form.is_valid():
         form.save()
         return redirect('articles:home')
@@ -47,8 +54,14 @@ def post_edit(request, pk):
         'post_edit.html',
         {'form': form, 'article': article})
 
+
+@login_required(login_url='/users/sign_in')
 def post_delete(request, pk):
-    article = Post.objects.get( pk=pk)
+    article = Post.objects.get(pk=pk)
+
+    if article.author != request.user:
+        return redirect('users:forbidden')
+
     if request.method == "POST":
         article.delete()
     return redirect('articles:home')
